requirements.txt
astroid==2.3.1
autopep8==1.4.4
backcall==0.1.0
blis==0.4.1
bokeh==1.4.0
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.3
dash==1.7.0
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
decorator==4.4.1
Flask==1.1.1
Flask-Compress==1.4.0
fr-core-news-sm==2.2.5
future==0.18.2
idna==2.8
importlib-metadata==1.2.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipython-sql==0.3.9
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
more-itertools==8.0.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.4
packaging==19.2
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
plotly==4.4.1
preshed==3.0.2
prettytable==0.7.2
prompt-toolkit==3.0.2
psycopg2==2.8.3
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.2
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.1
requests==2.22.0
retrying==1.3.3
six==1.12.0
spacy==2.2.3
SQLAlchemy==1.3.11
sqlparse==0.3.0
srsly==0.2.0
thinc==7.3.1
tornado==6.0.3
tqdm==4.40.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wasabi==0.4.2
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
